refactored workDir check and removed -f option
